Refactored AutocompleteCommandBox
